audio: Megre T3 P code to Android R  [1/1]
PD#SWPL-61871

Problem:
We need merge android P related code to Android R

Solution:
Megre T3 P audio rebase code to Android

Verify:
T3

+++ audio/aml_resampler/include/aml_audio_resample_manager.h
@@ -43,7 +43,7 @@ typedef struct aml_audio_resample {
     void *resample_handle;
     size_t total_in;
     size_t total_out;
-    int align_size;
+    size_t align_size;
     int last_left_size;
     int last_copy_size;
 } aml_audio_resample_t;
